arith and logical taking immediate values now
need to to memory, as described in a comment, and soon will have to do unit tests to test all of the instructions to make sure they are generating correctly. 
also need to make a "wrapper" of sorts to make it easier to type for the average user. will brainstorm and or talk to people about it.